article-fetcher/requirements.txt
fastapi
uvicorn
python-dotenv
news-please
httpx